+++ edfoil/classes/airfoil.py
@@ -123,8 +123,27 @@ class Airfoil:
             self.n_points = len(coords)
         
     def update(self,coords:list) -> None:
-        self.xy = coords
+        # self.xy = coords
+        # self.n_points = len(coords)
+        # Split at min x (leading edge)
+        # Big Assumption: Coordinates are ordered from TE to LE
+        coords = np.array(coords)
+        le_index = np.argmin(coords[:,0])
+        
+        side_1 = coords[:le_index+1]
+        side_2 = coords[le_index:]
+        
+        if np.mean(side_1[:,1]) >= np.mean(side_2[:,1]):
+            self.upper = side_1[side_1[:, 0].argsort()].tolist()
+            self.lower = side_2[side_2[:, 0].argsort()].tolist()
+        else:
+            self.upper = side_2[side_2[:, 0].argsort()].tolist()
+            self.lower = side_1[side_1[:, 0].argsort()].tolist()
+            
+        self.xy = self.lower[::-1] + self.upper[1:]
+        # self.xy = coords.tolist()
         self.n_points = len(coords)
+        
     
     def changeName(self, name:str) -> None:
         self.name = name